Update README.md
This is an inspirational movie of a poor guy achieving the dream of being a football player. 
Santiago was poor and he was a illegal immigrant from Mexico to USA, He helped his father in his work. He has a dream of being a foot ball player. He saved money in his shoe to buy football kit. His grandmother is also a great football player. He was so poor that he used pair of cardboard pieces as shin guards. 
One fine day, while he was playing the game, He was noticed by Glen Foy, he is an ex football player, Glen promises Santiago that he will set up a meeting with New castle manager. Glen tells Santiago that if he can go to England he will be given a chance to play for new castle united. His grandmother sells off part of her jewellery for his ticket from US to Mexico and from Mexico to London as he was a illegal immigrant. 
In his first match at England he didn’t play well. Glen manages to convince the manager saying that he was nervous and jetlagged. Santiago passed medical test by lying about his asthma. He uses a inhaler, which was crushed by a team-mate in a match, due to this he didn’t play further matches well and he was removed from the club.
On his way to airport he meets Harris in a Taxi, Gavin Harris was a part of new castle team. Harris tries to convince team manager to allow Santiago some more time in the new castle team. He gets confident and plays well in the game . He was a substitute in the match against Fulham. He helped team to win the match by a penalty goal. His weakness was that he doesn’t pass on the ball. Santiago transferred back to the reserve and he was demotivated and he gets a call that his father got expired, His heart broken but he decided not to go back but wanted to stay back and play the match. 
In the final minute Santiago pass on the ball to Harris, He corrected his draw back. Harris decided Santiago to kick the goal. He finally kicks the goal and New castle team wins the match. He finally achieves his dream. 

Learning’s:
1.	Even if you’re poor don’t stop dreaming big, Put your efforts.
2.	Try to be strong against all the odds, Don’t lose hope
3.	Be confident about your-self 
4.	Try to work on your weaknesses don’t get demotivated 
5.	 Believe in yourself one day you will achieve the impossible